Bindgen RSTRUCT_EMBED_LEN_MASK as usize

+++ zjit/src/cruby_bindings.inc.rs
@@ -2116,7 +2116,7 @@ pub struct rb_zjit_runtime_offsets {
 }
 pub const RSTRUCT_EMBED_LEN_MASK: ruby_rstruct_flags = 1040384;
 pub const RSTRUCT_EMBED_LEN_SHIFT: ruby_rstruct_flags = 13;
-pub type ruby_rstruct_flags = u32;
+pub type ruby_rstruct_flags = usize;
 pub const ROBJECT_OFFSET_AS_HEAP_FIELDS: jit_bindgen_constants = 16;
 pub const ROBJECT_OFFSET_AS_ARY: jit_bindgen_constants = 16;
 pub const RCLASS_OFFSET_PRIME_FIELDS_OBJ: jit_bindgen_constants = 40;
